backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.3.2
numpy==1.26.2
mcp==1.0.0
schedule==1.2.0
beautifulsoup4==4.12.2
